+++ R/bizdiff.R
@@ -18,7 +18,7 @@
 #'
 #' @examples
 #' dates <- c("2017-05-10", "2017-05-12", "2017-05-17")
-#' bizdiff(dates, "Brazil/ANBIMA")
+#' bizdiff(dates, "actual")
 #' @export
 bizdiff <- function(dates, cal) UseMethod("bizdiff")