+++ R/bizseq.R
@@ -18,10 +18,7 @@
 #' objects with ISO formatted dates.
 #'
 #' @examples
-#' create.calendar("Brazil/ANBIMA", holidaysANBIMA,
-#'   weekdays = c("saturday", "sunday")
-#' )
-#' bizseq("2013-01-02", "2013-01-31", "Brazil/ANBIMA")
+#' bizseq("2013-01-02", "2013-01-31", "actual")
 #' @export
 bizseq <- function(from, to, cal) UseMethod("bizseq")