+++ R/load-buildin-calendars.R
@@ -23,21 +23,8 @@
 #' bizdays::is.bizday("2020-01-01", "Brazil/ANBIMA")
 #' @export
 load_builtin_calendars <- function() {
-  create.calendar("actual")
-  localenv <- new.env()
-  utils::data("holidaysANBIMA", envir = localenv)
-  create.calendar("Brazil/ANBIMA", localenv$holidaysANBIMA,
-    weekdays = c("saturday", "sunday"),
-    adjust.from = adjust.next, adjust.to = adjust.previous
-  )
-  utils::data("holidaysB3", envir = localenv)
-  create.calendar("Brazil/B3", localenv$holidaysB3,
-    weekdays = c("saturday", "sunday"),
-    adjust.from = adjust.next, adjust.to = adjust.previous
-  )
-  create.calendar("weekends",
-    weekdays = c("saturday", "sunday"),
-    adjust.from = adjust.next, adjust.to = adjust.previous
-  )
+  cal_dir <- system.file("extdata", package = "bizdays")
+  fnames <- list.files(cal_dir, pattern = "json$", full.names = TRUE)
+  lapply(fnames, load_calendar)
   bizdays.options$set(default.calendar = "actual")
 }