+++ tests/testthat/test-getdate.R
@@ -1,0 +1,17 @@
+test_that("getdate", {
+  expect_equal(getdate("10th wed", 2018, "actual"), as.Date("2018-03-07"))
+  dts <- getdate("last bizday", 2010:2018, "Brazil/ANBIMA")
+  dts_1 <- as.Date(c(
+    "2010-12-31", "2011-12-30", "2012-12-31", "2013-12-31", "2014-12-31",
+    "2015-12-31", "2016-12-30", "2017-12-29", "2018-12-31"
+  ))
+  expect_equal(dts, dts_1)
+  dts <- seq(as.Date("2018-01-01"), as.Date("2018-12-01"), "month")
+  dts_1 <- getdate("first bizday", format(dts, "%Y-%m"), "Brazil/ANBIMA")
+  dts_2 <- as.Date(c(
+    "2018-01-02", "2018-02-01", "2018-03-01", "2018-04-02", "2018-05-02",
+    "2018-06-01", "2018-07-02", "2018-08-01", "2018-09-03", "2018-10-01",
+    "2018-11-01", "2018-12-03"
+  ))
+  expect_equal(dts_1, dts_2)
+})